requirements.txt
pytest
requests
allure-pytest
bestconfig
selenium